Renamemd files and bug fixed team projects
